seperating out the likes logic to a likes_controller
