+++ web/templates/homepage/show.html.eex
@@ -21,10 +21,10 @@
           <p class="b">Would you recommend this?</p>
           <div class="w-50 mw4 dib">
             <div class="dib mb3 v-mid tl"><span><%= resource.likes %></span></div>
-            <%= form_for @conn, homepage_path(@conn, :like, resource.id) <> "#resource_#{resource.id}", [as: :like, class: "dib w-30"], fn _f -> %>
+            <%= form_for @conn, likes_path(@conn, :like, resource.id) <> "#resource_#{resource.id}", [as: :like, class: "dib w-30"], fn _f -> %>
               <button title="like" type="submit" class="<%= check_liked "like", resource.liked %> w-100 h2 bn bg-white fl p0 bg-left"></button>
             <%= end %>
-            <%= form_for @conn, homepage_path(@conn, :dislike, resource.id) <> "#resource_#{resource.id}", [as: :dislike, class: "dib w-30"], fn _f -> %>
+            <%= form_for @conn, likes_path(@conn, :dislike, resource.id) <> "#resource_#{resource.id}", [as: :dislike, class: "dib w-30"], fn _f -> %>
               <button title="dislike" type="submit" class="<%= check_liked "dislike", resource.liked %> w-100 h2 bn bg-white fr p0 bg-right"></button>
             <%= end %>
             <div class="dib mb3 v-mid tl"><span><%= resource.dislikes %></span></div>
